Hep upgrade rework
Instead of upgrades from XL 1 to 27, there are now four distinct
upgrade points. The first comes at piety rank 2, when the player
chooses their ancestor's class. After that, at XL 15, 21, and 27,
there are further upgrades dependent on the class. The XL 15
upgrade offers a choice between two options, e.g. an axe or a
trident for the knight; subsequent upgrades are fixed. The goals
are to scale the ancestor to be useful in later parts of the game
and to allow the player to customize the ancestor to their own
playstyle, without overwhelming the player or the code with
needless complexity (as the old setup did).

Upgrades:
XL        Knight         |   Battlemage    |   Hexer
15a   D. Trident (flame) |   Iceblast      | Paralyse
15b   Broad Axe (flame)  | Bolt of Magma   | Metabol. Englaciation
21    Lg. Shield (refl.) | Lajatang (frz.) | Mass Confusion
      Haste                Haste             Haste
27    Speed brand        |  Crystal Spear  | Quickblade (AM)