Extra meliai vaults, mostly vaguely inspired by Hive
Not that these are necessarily dependent on melissae: they just obviously
benefit from the premise. Using the decent design precedent of the previous
_what_comes_after_a, these large vaults loosely based off old hive layouts
mix together and scatter floor spawns with killer bees in weird thick grids.

There's also a Shoals minivault, because they're as Greek as the rest of the
branch.