Experiment: buff jewellery acq power
Jewellery acquisition is generally considered Quite Bad, and a
skilled player will very rarely choose it. It seems like it'd
be more interesting if there were more contexts in which jewel
acq was a plausible choice.

I don't think that ?acq is 'overpowered' overall, so increasing
jewellery acq's power to try to compete with armour & weapon acq
seems like a reasonable approach. Hence, this ups the odds of
jewellery acq to produce artefacts (from ~30% to ~50%). Might
make it more competitive.

Or maybe we'd need the ability to specify 'less bad randarts' to
do that? Who knows. Worth a shot?